requirements.txt
black===23.9.1
flake8===6.1.0
isort==5.12.0
fastapi==0.97.0
uvicorn==0.22.0
pydantic==1.10.9
python-dotenv==1.0.0